requirements.txt
# Core AI Framework
crewai
crewai-tools
python-dotenv
litellm

# Web Framework (FastAPI Backend)
fastapi
uvicorn[standard]
websockets
pydantic
